Add debug caller library
Add the `debug-caller` library, with the basic functionality to create
a `debug` object with a decorator using the `caller` logic to determine
the calling filename.  Added some notes about how to do this a bit
easier, but for now, stick with the basic path.  The logger returns 2
logs: one for log, the other for error.  Also expose the `debug` module
in case the consumers need direct access.

Add tests for the basic functionality as well.