Add depth parameter, add docs/comments
Add the ability for the user to specify the depth used in generating
the calling filename.  This is useful when a consumer has wrapped
`debug-caller` in a central logging utility, and would like to proxy
requests through that logic.  In that case, they would specify “2” as
the depth rather than the default “1”.

Add additional documentation on the main library, and update the tests.