Options config, default no random colors
Change the configuration for `debug-caller` to take in the application
(or module) name as the first parameter, and the second is an options
object.  Within this options object, allow the `depth` parameter as
well as a new `randomColors` parameter which defaults to `false`.  If
`randomColors` is false, we assign white to the log and red to the
error debug instances returned to the user.

Add testing for these changes and update the documentation to explain.